# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "後端專案",
        "description": "\n *     後端專案專用的 ＡＰＩ 文件,\n *     請先透過 auth 分類中的 login 登入拿到 token\n *     在到右邊的 Authorize 按鈕中中設定，方可使用所有有鎖頭的 api\n *     然後有鎖頭的 ＡＰＩ 要在 http header 帶 token 進去\n *     ",
        "contact": {
            "name": "stephen chen",
            "email": "tasb00429@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/admins/authUser": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "把 Admins 資料列出來, 但是不會顯示當前的自己",
                "operationId": "b3c98ff4a5f79ba02af9bd784217c4e8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "admin 的 page ( 最小是 1 )",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "admin 的 per_page ( 默認 20 )",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "新增 Admin",
                "operationId": "ec1ae4f6306e708d6f7662e16776ebdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admins/authUser/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "查看一筆 Admin, 會把 role 一併回傳",
                "operationId": "3dd3feaad17383dc4ad7f9364cc39264",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "admin 的 id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "修改一筆 Admin",
                "operationId": "a36373c404a8a5f3f99b60abb7dd0167",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "admin 的 id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "刪除一筆 Admin",
                "operationId": "64e77ed97f5d70a854e03e7957e326c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "admin 的 id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admins/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "後台登入，測試帳密是 admin 或者 admin@gmail.com, 123456，登入之後請在打 me 這隻 api 拿到當前使用者資訊",
                "operationId": "5563155241b9391f950225221d545341",
                "parameters": [
                    {
                        "name": "primary",
                        "in": "query",
                        "description": "信箱 or 帳號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/admins/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "拿到 admin 使用者的相關資料",
                "operationId": "d43087e86b4255b9e9d8dac3c1be1bbb",
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admins/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log the user out (Invalidate the token)",
                "operationId": "909b1ba20f6c909f0a2706446d1d3ae6",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "信箱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admins/refresh": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh a token.",
                "operationId": "de1ed6aafc6a06d202ec7d7d197d6c86",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "信箱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AdminModel": {
                "title": "AdminModel",
                "required": [
                    "name",
                    "photoUrls"
                ],
                "properties": {
                    "account": {
                        "type": "string",
                        "format": "string",
                        "example": "account"
                    },
                    "email": {
                        "type": "string",
                        "format": "string",
                        "example": "xxxx@gamil.com"
                    },
                    "password": {
                        "description": "默認是驗證格式是 `數字` + `大寫英文` + `小寫英文` + `開頭是否是英文` + `下底線` + `橫線`",
                        "type": "string",
                        "format": "string",
                        "example": "password111"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "string",
                        "example": "password111"
                    },
                    "display_name": {
                        "type": "string",
                        "format": "string",
                        "example": "小虎"
                    },
                    "status": {
                        "description": "狀態，可能會有 0 ~ 127",
                        "type": "string",
                        "format": "integer",
                        "example": "0"
                    },
                    "role_id": {
                        "type": "string",
                        "format": "integer",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AdminModel"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}